requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
future==0.18.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
line-bot-sdk==1.16.0
mccabe==0.6.1
psycopg2==2.8.5
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.11.2
